% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_funcs.R
\name{ID_to_ENTREZ}
\alias{ID_to_ENTREZ}
\title{Convert input biological IDs to ENTREZ Ids}
\usage{
ID_to_ENTREZ(x, from)
}
\arguments{
\item{x}{vector of input IDs}

\item{from}{string to denote input ID type}
}
\value{
Vector of input IDs converted to ENTREZ IDs
}
\description{
For a list of acceptable input IDs and acceptable strings; use valid_ids()
}
