% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_funcs.R
\name{valid_input}
\alias{valid_input}
\title{Check valid input IDs and corresponding strings for converting to ENTREZ}
\usage{
valid_input()
}
\description{
Use one of the valid input IDs returned by this function to output ENTREZ IDs
for enrichment analysis
}
