% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_funcs.R
\name{run_enrichment}
\alias{run_enrichment}
\title{Run enrichment across the different pathways based on the the input gene set of interest and
background set}
\usage{
run_enrichment(x, y)
}
\arguments{
\item{x}{List of ENTREZ Ids of interest}

\item{y}{List of ENTREZ Ids of background set}
}
\value{
Dataframe of significant enrichment results at FDR< 0.05
}
\description{
Run enrichment across the different pathways based on the the input gene set of interest and
background set
}
